various small fixes, separate tracker_race_test
